refactor: Todo 타입 위치 변경
